refactor(constant): move file with constant to new directory
